modules/step-by-step-projects/projects/rag/ragproject-local-02-03-2026/pyproject.toml
[project]
name = "ragproject"
version = "0.1.0"
description = ""
authors = [
    {name = "rodrigols89",email = "rodrigols89.inbox@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"

dependencies = [
    "django (>=6.0.2,<7.0.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "cryptography (>=46.0.5,<47.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "django-allauth (>=65.14.3,<66.0.0)"
]

[dependency-groups]
dev = [
    "poetry-plugin-export (>=1.10.0,<2.0.0)"
]



#############################################################
#                     Taskipy Section                       #
#############################################################

[tool.taskipy.tasks]
# -------------- ( General Docker Management ) --------------
start_compose = 'docker compose up -d'
down_compose = 'docker compose down'
restart_compose = 'docker restart $(docker ps -q)'
build_compose = 'docker compose up --build -d'
clean_compose = """
docker stop $(docker ps -aq) 2>/dev/null || true &&
docker rm $(docker ps -aq) 2>/dev/null || true &&
docker rmi -f $(docker images -aq) 2>/dev/null || true &&
docker volume rm $(docker volume ls -q) 2>/dev/null || true &&
docker system prune -a --volumes -f
"""
# --------------------- ( Docker Exec ) ---------------------
opendb = 'docker exec -it database psql -U raguser -d rag_db'
openweb = 'docker exec -it django bash'
opennginx = 'docker exec -it nginx bash'
openredis = 'docker exec -it redis bash'
# ------------------ ( Project Management ) -----------------
createuser = 'docker compose exec web python manage.py createsuperuser'
migrate = 'docker compose exec web python manage.py migrate'
makemigrations = 'docker compose exec web python manage.py makemigrations'
exportdev = """
docker compose exec --user appuser web \
poetry export \
--without-hashes \
--with dev \
--format=requirements.txt \
--output=requirements-dev.txt
"""
exportprod = """
docker compose exec --user appuser web \
poetry export \
--without-hashes \
--format=requirements.txt \
--output=requirements.txt
"""
exportlocal = 'pip freeze > requirements-local.txt --require-virtualenv'
addpermissions = """
sudo chown -R 1000:1000 ./static ./media ./staticfiles || true &&
sudo chmod -R 755 ./static ./media ./staticfiles
"""



#############################################################
#                    Build System Section                   #
#############################################################

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
